requirements.txt
ollama
pandas
openpyxl
streamlit
plotly
pytesseract
PyMuPDF
xlsxwriter
mysql-connector-python
python-dotenv
streamlit_authenticato